refactor(beneficiary-service): Split beneficiary routes out of controller
Moves route/middleware declarations and response Zod schemas into
beneficiary.routes.ts; controller keeps only request-handling logic via
createBeneficiaryController(service). module.ts now creates the shared
DocumentedRouter and registers routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after (8 routes), lint clean, all 184 existing tests pass
(including the ownership-check coverage added in the prior
GET /beneficiaries/:id fix), and live smoke checks confirming route
ordering (list/summary routes not swallowed by /:id) is preserved.
