+++ apps/beneficiary-service/src/beneficiary/beneficiary.module.ts
@@ -1,15 +1,17 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { BeneficiaryRepository } from './beneficiary.repository';
 import { BeneficiaryService } from './beneficiary.service';
-import { createBeneficiaryRouter } from './beneficiary.controller';
+import { registerBeneficiaryRoutes } from './beneficiary.routes';
 
 /**
- * Composition root for the beneficiary feature: wires repository → service → router.
+ * Composition root for the beneficiary feature: wires repository → service → routes.
  * Replaces the former NestJS module + DI container.
  */
 export function createBeneficiaryModule(prisma: PrismaService): DocumentedRouter {
   const repository = new BeneficiaryRepository(prisma);
   const service = new BeneficiaryService(repository);
-  return createBeneficiaryRouter(service);
+  const doc = createDocumentedRouter();
+  registerBeneficiaryRoutes(doc, service);
+  return doc;
 }
